Codebase cleanup & Menu work
