+++ Hampage/Assets/Materials/Colors/CoinMat.mat
@@ -63,9 +63,9 @@ Material:
     - _DetailNormalMapScale: 1
     - _DstBlend: 0
     - _GlossMapScale: 1
-    - _Glossiness: 0.897
+    - _Glossiness: 0.678
     - _GlossyReflections: 1
-    - _Metallic: 0
+    - _Metallic: 0.913
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
