--- Hampage/Assets/Scripts/PauseMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fb44b44619d0cba47b89b5e8807e5787
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
