+++ Hampage/ProjectSettings/TagManager.asset
@@ -10,6 +10,7 @@ TagManager:
   - Enemy
   - Map
   - MinimapCamera
+  - GameManager
   layers:
   - Default
   - TransparentFX
@@ -24,8 +25,6 @@ TagManager:
   - Enemy
   - Coin
   - MapIcons
-  - 
-  - 
   - 
   - 
   -
